postgresstore: refactor
